+++ src/gpu_1/icp/icp.cu
@@ -9,10 +9,10 @@ namespace gpu_1::icp
     std::size_t icp_gpu(const matrix_host_t& M_host /*dst*/,
                         const matrix_host_t& P_host /*src*/,
                         matrix_host_t& newP_host,
-                        double& err,
+                        float& err,
                         bool verbose,
                         std::size_t max_iterations,
-                        double threshold,
+                        float threshold,
                         std::size_t power_iteration_simulations)
     {
         matrix_device_t M(M_host.size(), M_host[0].size());
@@ -64,7 +64,7 @@ namespace gpu_1::icp
         // Find Correspondences
         matrix_device_t Y(newP.rows_, newP.cols_);
 
-        double scaling_factor = 0.0; // s
+        float scaling_factor = 0.0; // s
         matrix_device_t rotation_matrix(3, 3); // R
         matrix_device_t translation_matrix(1, 3); // t
 
@@ -140,7 +140,7 @@ namespace gpu_1::icp
 
     bool find_alignment(const matrix_device_t& P,
                         const matrix_device_t& Y,
-                        double& s,
+                        float& s,
                         matrix_device_t& R,
                         matrix_device_t& t,
                         std::size_t power_iteration_simulations)
@@ -193,33 +193,33 @@ namespace gpu_1::icp
 
         matrix_device_t xx(1, Pprime_T.cols_);
         utils::vector_element_wise_multiplication(Pprime_T, 0, Yprime_T, 0, xx);
-        double Sxx = utils::vector_sum(xx);
+        float Sxx = utils::vector_sum(xx);
         matrix_device_t xy(1, Pprime_T.cols_);
         utils::vector_element_wise_multiplication(Pprime_T, 0, Yprime_T, 1, xy);
-        double Sxy = utils::vector_sum(xy);
+        float Sxy = utils::vector_sum(xy);
         matrix_device_t xz(1, Pprime_T.cols_);
         utils::vector_element_wise_multiplication(Pprime_T, 0, Yprime_T, 2, xz);
-        double Sxz = utils::vector_sum(xz);
+        float Sxz = utils::vector_sum(xz);
 
         matrix_device_t yx(1, Pprime_T.cols_);
         utils::vector_element_wise_multiplication(Pprime_T, 1, Yprime_T, 0, yx);
-        double Syx = utils::vector_sum(yx);
+        float Syx = utils::vector_sum(yx);
         matrix_device_t yy(1, Pprime_T.cols_);
         utils::vector_element_wise_multiplication(Pprime_T, 1, Yprime_T, 1, yy);
-        double Syy = utils::vector_sum(yy);
+        float Syy = utils::vector_sum(yy);
         matrix_device_t yz(1, Pprime_T.cols_);
         utils::vector_element_wise_multiplication(Pprime_T, 1, Yprime_T, 2, yz);
-        double Syz = utils::vector_sum(yz);
+        float Syz = utils::vector_sum(yz);
 
         matrix_device_t zx(1, Pprime_T.cols_);
         utils::vector_element_wise_multiplication(Pprime_T, 2, Yprime_T, 0, zx);
-        double Szx = utils::vector_sum(zx);
+        float Szx = utils::vector_sum(zx);
         matrix_device_t zy(1, Pprime_T.cols_);
         utils::vector_element_wise_multiplication(Pprime_T, 2, Yprime_T, 1, zy);
-        double Szy = utils::vector_sum(zy);
+        float Szy = utils::vector_sum(zy);
         matrix_device_t zz(1, Pprime_T.cols_);
         utils::vector_element_wise_multiplication(Pprime_T, 2, Yprime_T, 2, zz);
-        double Szz = utils::vector_sum(zz);
+        float Szz = utils::vector_sum(zz);
 
         matrix_device_t Nmatrix(4, 4);
         Nmatrix.set_val(0, 0, Sxx + Syy + Szz);
@@ -258,8 +258,8 @@ namespace gpu_1::icp
 
         // ----------------------------------------
         // Scaling factor computation
-        double Sp = 0.0;
-        double D = 0.0;
+        float Sp = 0.0;
+        float D = 0.0;
 
         matrix_device_t dot_product(1, 1);
 
@@ -318,7 +318,7 @@ namespace gpu_1::icp
     void power_iteration(const matrix_device_t& A, matrix_device_t& eigen_vector, std::size_t num_simulations)
     {
         vector_host_t vector(A.cols_);
-        std::generate_n(vector.begin(), A.cols_, utils::UniformRandom<double>(0.0, 1.1));
+        std::generate_n(vector.begin(), A.cols_, utils::UniformRandom<float>(0.0, 1.1));
         for (std::size_t i = 0; i < A.cols_; i++)
         {
             eigen_vector.set_val(i, 0, vector[i]);
@@ -330,7 +330,7 @@ namespace gpu_1::icp
         {
             utils::matrix_dot_product(A, eigen_vector, b_k1);
 
-            double b_k1_norm = b_k1.matrix_norm_2();
+            float b_k1_norm = b_k1.matrix_norm_2();
 
             for (std::size_t i = 0; i < eigen_vector.rows_; i++)
             {
@@ -340,7 +340,7 @@ namespace gpu_1::icp
     }
 
     void apply_alignment(const matrix_device_t& P,
-                         double s,
+                         float s,
                          const matrix_device_t& R,
                          const matrix_device_t& t,
                          matrix_device_t& newP)
