+++ src/gpu_1/parser/parser.hh
@@ -8,7 +8,7 @@
 
 namespace gpu_1::parser
 {
-    using value_t = double;
+    using value_t = float;
     using vector_host_t = std::vector<value_t>;
     using matrix_host_t = std::vector<vector_host_t>;
